+++ src/elements/actions/GenerateAiAltText.php
@@ -6,10 +6,7 @@ use Craft;
 use craft\base\ElementAction;
 use craft\elements\Asset;
 use craft\elements\db\ElementQueryInterface;
-use craft\helpers\ElementHelper;
-use craft\helpers\Queue;
 use heavymetalavo\craftaialttext\AiAltText;
-use heavymetalavo\craftaialttext\jobs\GenerateAiAltText as GenerateAiAltTextJob;
 use yii\base\InvalidConfigException;
 
 /**
@@ -62,67 +59,16 @@ class GenerateAiAltText extends ElementAction
             throw new InvalidConfigException('User not logged in');
         }
 
-        $elementsService = Craft::$app->getElements();
-        $queue = Craft::$app->getQueue();
-
-        foreach ($query->all() as $element) {
-            if (!$element instanceof Asset) {
+        foreach ($query->all() as $asset) {
+            if (!$asset instanceof Asset) {
                 continue;
             }
 
-            // Check if there's already a job for this element
-            $existingJobs = $queue->getJobInfo();
-            $hasExistingJob = false;
-            foreach ($existingJobs as $job) {
-                if (isset($job['description']) && strpos($job['description'], "Element: {$element->id}") !== false) {
-                    $hasExistingJob = true;
-                    break;
-                }
-            }
+            // Set the current site id on asset
+            $asset = Asset::find()->id($asset->id)->siteId($query->siteId)->one();
 
-            if ($hasExistingJob) {
-                Craft::$app->getSession()->setNotice(Craft::t('ai-alt-text', "{$element->filename} (ID: {$element->id}) is already being processed within an existing queued job. Please wait for the existing job to finish before attempting to process it again."));
-                continue;
-            }
-
-            if ($element->kind !== Asset::KIND_IMAGE) {
-                continue;
-            }
-
-            $saveTranslatedResultsToEachSite = AiAltText::getInstance()->settings->saveTranslatedResultsToEachSite;
-
-            // Queue a job for the current site
-            $queue->push(new GenerateAiAltTextJob([
-                'description' => Craft::t('ai-alt-text', 'Generating alt text for {filename} (Element: {id}, Site: {siteId})', [
-                    'filename' => $element->filename,
-                    'id' => $element->id,
-                    'siteId' => $element->siteId,
-                ]),
-                'elementId' => $element->id,
-                'siteId' => $element->siteId,
-            ]));
-
-            // If we're saving results to each site and translated results for each site, we need to queue a job for each site
-            if ($saveTranslatedResultsToEachSite) {
-                foreach (Craft::$app->getSites()->getAllSites() as $site) {
-                    // Skip the current site
-                    if ($site->id === $element->siteId) {
-                        continue;
-                    }
-
-                    $queue->push(new GenerateAiAltTextJob([
-                        'description' => Craft::t('ai-alt-text', 'Generating alt text for {filename} (Element: {id}, Site: {siteId})', [
-                            'filename' => $element->filename,
-                            'id' => $element->id,
-                            'siteId' => $site->id,
-                        ]),
-                        'elementId' => $element->id,
-                        'siteId' => $site->id,
-                    ]));
-                }
-            }
-            
-
+            // Create a job for the asset
+            AiAltText::getInstance()->aiAltTextService->createJob($asset, true);
         }
 
         return true;