+++ src/services/AiAltTextService.php
@@ -5,13 +5,8 @@ namespace heavymetalavo\craftaialttext\services;
 use Craft;
 use craft\base\Component;
 use craft\elements\Asset;
-use craft\helpers\Assets;
-use craft\helpers\ElementHelper;
-use craft\helpers\Html;
-use craft\helpers\Template;
-use craft\web\View;
 use heavymetalavo\craftaialttext\AiAltText;
-use craft\helpers\App;
+use heavymetalavo\craftaialttext\jobs\GenerateAiAltText as GenerateAiAltTextJob;
 use Exception;
 use craft\events\DefineMenuItemsEvent;
 use craft\enums\MenuItemType;
@@ -40,6 +35,79 @@ class AiAltTextService extends Component
     }
 
     /**
+     * Creates a job for the given element
+     */
+    public function createJob(Asset $asset, $saveCurrentSiteOffQueue = false): void
+    {
+        $queue = Craft::$app->getQueue();
+        // Check if there's already a job for this element
+        $existingJobs = $queue->getJobInfo();
+        $hasExistingJob = false;
+        foreach ($existingJobs as $job) {
+            if (isset($job['description']) && str_contains($job['description'], "Asset: $asset->id")) {
+                $hasExistingJob = true;
+                break;
+            }
+        }
+
+        if ($hasExistingJob) {
+            Craft::$app->getSession()->setNotice(Craft::t('ai-alt-text', "$asset->filename (ID: $asset->id) is already being processed within an existing queued job. Please wait for the existing job to finish before attempting to process it again."));
+            return;
+        }
+
+        if ($asset->kind !== Asset::KIND_IMAGE) {
+            Craft::$app->getSession()->setNotice(Craft::t('ai-alt-text', "$asset->filename (ID: $asset->id) is not an image"));
+            return;
+        }
+
+        // Get the $saveTranslatedResultsToEachSite setting value
+        $saveTranslatedResultsToEachSite = AiAltText::getInstance()->settings->saveTranslatedResultsToEachSite;;
+
+        // Check if we need to save the current site off queue
+        if ($saveCurrentSiteOffQueue) {
+            $this->generateAltText($asset, $asset->siteId);
+    
+            if (!$saveTranslatedResultsToEachSite) {
+                return;
+            }
+        }
+
+        // Save the current site on queue
+        $queue->push(new GenerateAiAltTextJob([
+            'description' => Craft::t('ai-alt-text', 'Generating alt text for {filename} (Asset: {id}, Site: {siteId})', [
+                'filename' => $asset->filename,
+                'id' => $asset->id,
+                'siteId' => $asset->siteId,
+            ]),
+            'assetId' => $asset->id,
+            'siteId' => $asset->siteId,
+        ]));
+
+        // return early if we're not saving translated results to each site
+        if (!$saveTranslatedResultsToEachSite) {
+            return;
+        }
+
+        // If we're saving results to each site and translated results for each site, we need to queue a job for each site
+        foreach (Craft::$app->getSites()->getAllSites() as $site) {
+            // Skip the current site
+            if ($saveCurrentSiteOffQueue && $site->id === $asset->siteId) {
+                continue;
+            }
+
+            $queue->push(new GenerateAiAltTextJob([
+                'description' => Craft::t('ai-alt-text', 'Generating alt text for {filename} (Asset: {id}, Site: {siteId})', [
+                    'filename' => $asset->filename,
+                    'id' => $asset->id,
+                    'siteId' => $site->id,
+                ]),
+                'assetId' => $asset->id,
+                'siteId' => $site->id,
+            ]));
+        }
+    }
+
+    /**
      * Generates alt text for an asset using AI.
      *
      * This method:
@@ -53,10 +121,6 @@ class AiAltTextService extends Component
      */
     public function generateAltText(Asset $asset, int $siteId = null): string
     {
-        if (!$asset) {
-            throw new Exception('Asset cannot be null');
-        }
-
         if ($asset->kind !== Asset::KIND_IMAGE) {
             throw new Exception('Asset must be an image');
         }
@@ -75,44 +139,13 @@ class AiAltTextService extends Component
         }
 
         $asset->alt = $altText;
-        $runValidation = true;
-        if (!Craft::$app->elements->saveElement($asset, $runValidation)) {
+        if (!Craft::$app->elements->saveElement($asset, true)) {
             throw new Exception('Failed to save alt text for asset: ' . $asset->filename);
         }
 
         Craft::info('Successfully saved alt text for asset: ' . $asset->filename, __METHOD__);
         return $altText;
 
-    }
-
-    /**
-     * Validates an asset for alt text generation.
-     *
-     * This method checks that:
-     * - The asset is not null
-     * - The asset is an image
-     * - The asset has either a public URL or is accessible via file system
-     *
-     * @param Asset $asset The asset to validate
-     * @return bool True if the asset is valid, false otherwise
-     * @throws Exception If the asset is invalid
-     */
-    private function validateAsset(Asset $asset): bool
-    {
-        if (!$asset) {
-            throw new Exception('Asset cannot be null');
-        }
-
-        if ($asset->kind !== Asset::KIND_IMAGE) {
-            throw new Exception('Asset must be an image');
-        }
-
-        // Check for either public URL or file system access
-        if (!$asset->getUrl() && !$asset->getPath()) {
-            throw new Exception('Asset must have either a public URL or be accessible via file system');
-        }
-
-        return true;
     }
 
     /**